# pkg/openapi3/types.gen.go
// Package openapi3 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.0 DO NOT EDIT.
package openapi3

// ErrorLog defines model for ErrorLog.
type ErrorLog struct {
	Query     *string `json:"query"`
	RootCause *string `json:"rootCause"`
	Scope     *string `json:"scope"`
	Status    *string `json:"status"`
	Trace     *string `json:"trace"`
}

// ErrorLogs defines model for ErrorLogs.
type ErrorLogs = []ErrorLog

// Message defines model for Message.
type Message struct {
	Count     *string    `json:"count"`
	ErrorLog  *ErrorLogs `json:"errorLog,omitempty"`
	HostName  *string    `json:"hostName"`
	Status    *string    `json:"status"`
	TimeTaken *string    `json:"timeTaken"`
}

// CompetitionCrawlResponse defines model for CompetitionCrawlResponse.
type CompetitionCrawlResponse struct {
	CompetitionName *string  `json:"competitionName"`
	Message         *Message `json:"message,omitempty"`
}

// CompetitionCrawlRequest defines model for CompetitionCrawlRequest.
type CompetitionCrawlRequest struct {
	CompetitionID *string `json:"competitionID,omitempty"`
	Date          *string `json:"date,omitempty"`
}

// CrawlCompJSONBody defines parameters for CrawlComp.
type CrawlCompJSONBody struct {
	CompetitionID *string `json:"competitionID,omitempty"`
	Date          *string `json:"date,omitempty"`
}

// CrawlCompJSONRequestBody defines body for CrawlComp for application/json ContentType.
type CrawlCompJSONRequestBody CrawlCompJSONBody