# swagger-ui/openapi3.yaml
components:
  requestBodies:
    CompetitionCrawlRequest:
      content:
        application/json:
          schema:
            properties:
              competitionID:
                minLength: 1
                type: string
              date:
                minLength: 1
                type: string
      description: Request used for crawling competition data
      required: true
  responses:
    CompetitionCrawlResponse:
      content:
        application/json:
          schema:
            nullable: true
            properties:
              competitionName:
                nullable: true
                type: string
              message:
                $ref: '#/components/schemas/Message'
      description: Response with competition crawl results
  schemas:
    ErrorLog:
      properties:
        query:
          nullable: true
          type: string
        rootCause:
          nullable: true
          type: string
        scope:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
        trace:
          nullable: true
          type: string
      type: object
    ErrorLogs:
      items:
        $ref: '#/components/schemas/ErrorLog'
      type: array
    Message:
      properties:
        count:
          nullable: true
          type: string
        errorLog:
          $ref: '#/components/schemas/ErrorLogs'
        hostName:
          nullable: true
          type: string
        status:
          nullable: true
          type: string
        timeTaken:
          nullable: true
          type: string
      type: object
info:
  contact:
    url: https://github.com/CalebTracey/rugby-crawler-api
  description: REST API rugby data web crawler
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Rugby Data Crawler REST API
  version: 0.0.0
openapi: 3.0.0
paths:
  /competition:
    description: Crawl Competition
    post:
      operationId: CrawlComp
      requestBody:
        $ref: '#/components/requestBodies/CompetitionCrawlRequest'
      responses:
        "201":
          $ref: '#/components/responses/CompetitionCrawlResponse'
        "400":
          $ref: '#/components/responses/CompetitionCrawlResponse'
        "500":
          $ref: '#/components/responses/CompetitionCrawlResponse'
    summary: Crawl Requests
servers:
- description: Local development
  url: http://0.0.0.0:6080